restructure files / fix nits
