+++ packages/integration-testing/src/test/CliServiceLauncher.test.ts
@@ -20,11 +20,14 @@ import { join } from "node:path";
 
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 
-import { CliServiceLauncher } from "../CliServiceLauncher.js";
-import type { FoundryCliService, ServiceState } from "../FoundryCliService.js";
+import { CliServiceLauncher } from "../cli-service/CliServiceLauncher.js";
+import type {
+  FoundryCliService,
+  ServiceState,
+} from "../cli-service/FoundryCliService.js";
+import { OntologyServer } from "../cli-service/OntologyServer.js";
+import { StatusServer } from "../cli-service/StatusServer.js";
 import type { ServiceName } from "../generated/cli/index.js";
-import { OntologyServer } from "../OntologyServer.js";
-import { StatusServer } from "../StatusServer.js";
 import { NoSpawnStatusServer } from "./NoSpawnStatusServer.js";
 import { StubService } from "./StubService.js";
 import {
